Reorganized Q&A content and metadata
Placed key metadata in the content file front matter so items like the
images and description don't need to be hard coded in Markdown with Hugo
shortcodes. This makes the separation of content and design much better
so that we can more easily change the design in the future. Also means
that the RSS generated doesn't include those extra elements.