Conditionally set article:modified_time for /latest/
lastmod is set based on the git stamp, but when /latest/ is auto
generated each week, there is no new commit, so the date will show as
last modified when the page was actually committed to. This sets a
condition for /latest/ to pull the last mod from the data files, then
fallback to the lastmod param used on every other page.