Removed box-collapsible partial
I removed this and merged the code back into the higher level template
(betas/single.html) because it was becoming complex with the nesting.
box-collapsible was then calling another partial (subscribe), which I
couldn't figure out how to cascade the dot context down to. This is needed
to replace the RelPermaLink into the REFERRER field of the subscriber
form so we can track which page subscribers join from. Although this
duplicates some code in this betas single template, it makes it easier
to understand what's going on without passing lots of variables through
the dict, and also means we only need to send the dot context through to
one level of partial.