Adjusted table of contents
Google is parsing out the <ul> in a way that doesn't make sense for the search results. This is because we built the TOC for humans, so the items like "Also considered" make sense for the TOC, but not a Google SERP. This adjusts the menu so it would make more sense if it was parsed out, but ideally the "Dynatrace..." and "Nagios..." lines wouldn't show up. Next will be to try using sub-UL for them.