Adjusted side menu to 8 items
The Google snippet disappeared with the previous change that adjusted the structure of the menu. Theory is that this is because it has fewer tools listed. This adds in sub items into the menu below the 4 recommended tools, picking out some tools we also considered so there are 8 in total.