TOC highlighting fix
Fixes TOC issue here https://3.basecamp.com/4477159/buckets/19757286/todos/4237346749, now no block is skipped from highlighting when scrolling.

There is a little 'unselected' gap when crossing sections, but in order to fix that we should turn review all of the article section spacing which is a big job. This fixes at least the bigger functional issue of skipping an entire section highlighting.