Move Hugo config to config/ dir
Now we're using minified assets, the correct base URL is needed to allow
for deployments of different files to test and prod. Base URL must be
specified otherwise the URL is wrong. You can't use a relative URL for
the canonical tag, for example.

This splits the configs into a default and production directory. Default
will always use the prod base URL and test uses the test environment,
with the --environment flag set for builds.